+++ relay-server/src/metrics_extraction/transactions.rs
@@ -5,7 +5,7 @@ use std::collections::BTreeSet;
 use {
     relay_common::UnixTimestamp,
     relay_general::protocol::{AsPair, Event, EventType},
-    relay_general::store::normalize_dist,
+    relay_general::store::{get_breakdown_measurements, normalize_dist, BreakdownsConfig},
     relay_metrics::{Metric, MetricUnit, MetricValue},
     std::collections::BTreeMap,
     std::fmt::Write,
@@ -60,24 +60,28 @@ fn extract_dist(transaction: &Event) -> Option<String> {
 #[cfg(feature = "processing")]
 pub fn extract_transaction_metrics(
     config: &TransactionMetricsConfig,
+    breakdowns_config: Option<&BreakdownsConfig>,
     event: &Event,
     target: &mut Vec<Metric>,
-) {
+) -> bool {
     use relay_metrics::DurationPrecision;
 
     use crate::metrics_extraction::utils::with_tag;
 
     if event.ty.value() != Some(&EventType::Transaction) {
-        return;
+        return false;
     }
 
     if config.extract_metrics.is_empty() {
-        return;
+        relay_log::trace!("dropping all transaction metrics because of empty allow-list");
+        return false;
     }
 
     let mut push_metric = move |metric: Metric| {
         if config.extract_metrics.contains(&metric.name) {
             target.push(metric);
+        } else {
+            relay_log::trace!("dropping metric {} because of allow-list", metric.name);
         }
     };
 
@@ -91,7 +95,7 @@ pub fn extract_transaction_metrics(
         .and_then(|ts| UnixTimestamp::from_datetime(ts.into_inner()))
     {
         Some(ts) => ts,
-        None => return,
+        None => return false,
     };
 
     let mut tags = BTreeMap::new();
@@ -149,21 +153,16 @@ pub fn extract_transaction_metrics(
     }
 
     // Breakdowns
-    if let Some(breakdowns) = event.breakdowns.value() {
-        for (breakdown, annotated) in breakdowns.iter() {
-            let measurements = match annotated.value() {
-                Some(measurements) => measurements,
-                None => continue,
-            };
-
-            for (name, annotated) in measurements.iter() {
+    if let Some(breakdowns_config) = breakdowns_config {
+        for (breakdown_name, measurements) in get_breakdown_measurements(event, breakdowns_config) {
+            for (measurement_name, annotated) in measurements.iter() {
                 let measurement = match annotated.value().and_then(|m| m.value.value()) {
                     Some(measurement) => *measurement,
                     None => continue,
                 };
 
                 push_metric(Metric {
-                    name: metric_name(&["breakdowns", breakdown, name]),
+                    name: metric_name(&["breakdowns", breakdown_name, measurement_name]),
                     unit: MetricUnit::None,
                     value: MetricValue::Distribution(measurement),
                     timestamp,
@@ -198,6 +197,8 @@ pub fn extract_transaction_metrics(
             None => tags.clone(),
         },
     });
+
+    true
 }
 
 #[cfg(feature = "processing")]
@@ -226,6 +227,8 @@ fn get_measurement_rating(name: &str, value: f64) -> Option<String> {
 #[cfg(feature = "processing")]
 mod tests {
     use super::*;
+
+    use relay_general::store::BreakdownsConfig;
     use relay_general::types::Annotated;
     use relay_metrics::DurationPrecision;
 
@@ -235,6 +238,7 @@ mod tests {
         {
             "type": "transaction",
             "timestamp": "2021-04-26T08:00:00+0100",
+            "start_timestamp": "2021-04-26T07:59:01+0100",
             "release": "1.2.3",
             "dist": "foo ",
             "environment": "fake_environment",
@@ -247,24 +251,38 @@ mod tests {
                 "foo": {"value": 420.69},
                 "lcp": {"value": 3000.0}
             },
-            "breakdowns": {
-                "breakdown1": {
-                    "bar": {"value": 123.4}
-                },
-                "breakdown2": {
-                    "baz": {"value": 123.4},
-                    "zap": {"value": 666},
-                    "zippityzoppity": {"value": 666}
-                }
-            }
+            "spans": [
+                {
+                    "description": "<OrganizationContext>",
+                    "op": "react.mount",
+                    "parent_span_id": "8f5a2b8768cafb4e",
+                    "span_id": "bd429c44b67a3eb4",
+                    "start_timestamp": 1597976393.4619668,
+                    "timestamp": 1597976393.4718769,
+                    "trace_id": "ff62a8b040f340bda5d830223def1d81"
+                }   
+            ]
         }
         "#;
+
+        let breakdowns_config: BreakdownsConfig = serde_json::from_str(
+            r#"
+            {
+                "span_ops": {
+                    "type": "spanOperations",
+                    "matches": ["react.mount"]
+                }
+            }
+        "#,
+        )
+        .unwrap();
 
         let event = Annotated::from_json(json).unwrap();
 
         let mut metrics = vec![];
         extract_transaction_metrics(
             &TransactionMetricsConfig::default(),
+            Some(&breakdowns_config),
             event.value().unwrap(),
             &mut metrics,
         );
@@ -276,9 +294,7 @@ mod tests {
             "extractMetrics": [
                 "sentry.transactions.measurements.foo",
                 "sentry.transactions.measurements.lcp",
-                "sentry.transactions.breakdowns.breakdown1.bar",
-                "sentry.transactions.breakdowns.breakdown2.baz",
-                "sentry.transactions.breakdowns.breakdown2.zap",
+                "sentry.transactions.breakdowns.span_ops.ops.react.mount",
                 "sentry.transactions.transaction.duration"
             ],
             "extractCustomTags": ["fOO"]
@@ -288,32 +304,29 @@ mod tests {
         .unwrap();
 
         let mut metrics = vec![];
-        extract_transaction_metrics(&config, event.value().unwrap(), &mut metrics);
+        extract_transaction_metrics(
+            &config,
+            Some(&breakdowns_config),
+            event.value().unwrap(),
+            &mut metrics,
+        );
 
-        assert_eq!(metrics.len(), 6);
+        assert_eq!(metrics.len(), 4, "{:?}", metrics);
 
         assert_eq!(metrics[0].name, "sentry.transactions.measurements.foo");
         assert_eq!(metrics[1].name, "sentry.transactions.measurements.lcp");
         assert_eq!(
             metrics[2].name,
-            "sentry.transactions.breakdowns.breakdown1.bar"
-        );
-        assert_eq!(
-            metrics[3].name,
-            "sentry.transactions.breakdowns.breakdown2.baz"
-        );
-        assert_eq!(
-            metrics[4].name,
-            "sentry.transactions.breakdowns.breakdown2.zap"
+            "sentry.transactions.breakdowns.span_ops.ops.react.mount"
         );
 
-        let duration_metric = &metrics[5];
+        let duration_metric = &metrics[3];
         assert_eq!(
             duration_metric.name,
             "sentry.transactions.transaction.duration"
         );
         if let MetricValue::Distribution(value) = duration_metric.value {
-            assert_eq!(value, 0.0);
+            assert_eq!(value, 59000.0);
         } else {
             panic!(); // Duration must be set
         }
@@ -362,7 +375,7 @@ mod tests {
         )
         .unwrap();
         let mut metrics = vec![];
-        extract_transaction_metrics(&config, event.value().unwrap(), &mut metrics);
+        extract_transaction_metrics(&config, None, event.value().unwrap(), &mut metrics);
 
         assert_eq!(metrics.len(), 1);